requirements.txt
# AI Engineering RAG - 依赖

# LangChain
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-experimental>=0.2.0
langchain-text-splitters>=1.1.0
langchain-classic>=0.1.0

# 向量数据库
chromadb>=0.4.0

# Embedding模型
sentence-transformers>=2.2.0

# 其他
python-dotenv>=1.0.0